cache the built project in the docker image
Having to rebuild the binary every time is kind of silly, though it does
allow you to run other analyses on the codebase even if the compilation
fails. I don't think that's quite as useful as not having to recompile
every time you want to run the binary via docker though.