use `MIT OR Apache-2.0` for everything by default
And remove the option for `AGPL-3.0-only`. This decision is based on
a conversation I had with people in the `#rust:matrix.org` room, during
which [this link][0] was shared. It's likely that a lot of places have
this policy, Google is just nice enough to state it publicly.

If people using this template still want to use `AGPL-3.0-only`, they
are free to do so, though I think `MIT OR Apache-2.0` is a better
default.

[0]: https://opensource.google/documentation/reference/using/agpl-policy