remove cargo-outdated from CI
This is a nice developer tool but realistically there are better ways to
do this particular task in an automated fashion. Something like
dependabot, for example. I haven't bothered looking for a GitLab
equivalent yet though. Either way, this can slow down CI runs pretty
significantly and it's not worth the cost to iteration time.