change how toolchain components are depended upon
The motivation for this is the `llvm-tools-preview` component, which
needs to live in the same derivation as `rustc`. If it's not done this
way, then tools parsing `rustc --print sysroot` (which most things do in
order to do this kind of thing) won't be able to find the tools they're
looking for.

This commit changes the obvious way to add components to the toolchain
in a manner that allows the aforementioned situation to function
properly, instead of failing for surprising reasons.